Review update on April 2018 -> make naming meaningfull and group controller/action
