+++ DBScript/mvc5_izenda.sql
@@ -2159,8 +2159,8 @@ INSERT [dbo].[IzendaSystemSetting] ([Id], [Name], [Value], [Version], [Deleted],
 INSERT [dbo].[IzendaSystemSetting] ([Id], [Name], [Value], [Version], [Deleted], [Created], [CreatedBy], [Modified], [ModifiedBy]) VALUES (N'e834da5d-6a19-485b-95e1-2190df0f545c', N'HelpSystemUrl', N'http://www.izenda.com', NULL, 0, NULL, NULL, NULL, NULL)
 INSERT [dbo].[IzendaSystemSetting] ([Id], [Name], [Value], [Version], [Deleted], [Created], [CreatedBy], [Modified], [ModifiedBy]) VALUES (N'e834da5d-6a19-485b-95e1-2190df0f545d', N'TokenTimeout', N'20', NULL, 0, NULL, NULL, NULL, NULL)
 INSERT [dbo].[IzendaSystemSetting] ([Id], [Name], [Value], [Version], [Deleted], [Created], [CreatedBy], [Modified], [ModifiedBy]) VALUES (N'463f5f94-9895-490e-8a7c-611c9487262c', N'DeploymentMode', N'1', NULL, 0, NULL, NULL, NULL, NULL)
-INSERT [dbo].[IzendaSystemSetting] ([Id], [Name], [Value], [Version], [Deleted], [Created], [CreatedBy], [Modified], [ModifiedBy]) VALUES (N'463f5f94-9895-490e-8a7c-611c9487262e', N'AuthValidateAccessTokenUrl', N'http://localhost:14809/api/account/validateIzendaAuthToken', NULL, 0, NULL, NULL, NULL, NULL)
-INSERT [dbo].[IzendaSystemSetting] ([Id], [Name], [Value], [Version], [Deleted], [Created], [CreatedBy], [Modified], [ModifiedBy]) VALUES (N'463f5f94-9895-490e-8a7c-611c9487262f', N'AuthGetAccessTokenUrl', N'http://localhost:14809/api/account/GetIzendaAccessToken', NULL, 0, NULL, NULL, NULL, NULL)
+INSERT [dbo].[IzendaSystemSetting] ([Id], [Name], [Value], [Version], [Deleted], [Created], [CreatedBy], [Modified], [ModifiedBy]) VALUES (N'463f5f94-9895-490e-8a7c-611c9487262e', N'AuthValidateAccessTokenUrl', N'http://localhost:14809/api/IzendaAccessToken/ValidateIzendaAuthToken', NULL, 0, NULL, NULL, NULL, NULL)
+INSERT [dbo].[IzendaSystemSetting] ([Id], [Name], [Value], [Version], [Deleted], [Created], [CreatedBy], [Modified], [ModifiedBy]) VALUES (N'463f5f94-9895-490e-8a7c-611c9487262f', N'AuthGetAccessTokenUrl', N'http://localhost:14809/api/IzendaAccessToken/GetIzendaAccessToken', NULL, 0, NULL, NULL, NULL, NULL)
 INSERT [dbo].[IzendaSystemSetting] ([Id], [Name], [Value], [Version], [Deleted], [Created], [CreatedBy], [Modified], [ModifiedBy]) VALUES (N'bb408fec-ec4b-4ef3-bbe2-702a2faa9c64', N'LicenseExpiredWarningPeriod', N'14', NULL, 0, NULL, NULL, NULL, NULL)
 INSERT [dbo].[IzendaSystemSetting] ([Id], [Name], [Value], [Version], [Deleted], [Created], [CreatedBy], [Modified], [ModifiedBy]) VALUES (N'6ed4ed91-7c3e-4200-ad2c-77d7f25a139a', N'CommonFilterSetting', N'0', NULL, 0, NULL, NULL, NULL, NULL)
 INSERT [dbo].[IzendaSystemSetting] ([Id], [Name], [Value], [Version], [Deleted], [Created], [CreatedBy], [Modified], [ModifiedBy]) VALUES (N'6ed4ed91-7c3e-4200-ad2c-77d7f25a139b', N'TracingType', N'3', NULL, 0, NULL, NULL, NULL, NULL)
@@ -2169,7 +2169,7 @@ INSERT [dbo].[IzendaSystemSetting] ([Id], [Name], [Value], [Version], [Deleted],
 INSERT [dbo].[IzendaSystemSetting] ([Id], [Name], [Value], [Version], [Deleted], [Created], [CreatedBy], [Modified], [ModifiedBy]) VALUES (N'b08778b0-28a2-449c-a859-8966a34cb45d', N'WebUrl', N'', NULL, 0, NULL, NULL, NULL, NULL)
 INSERT [dbo].[IzendaSystemSetting] ([Id], [Name], [Value], [Version], [Deleted], [Created], [CreatedBy], [Modified], [ModifiedBy]) VALUES (N'8d721bfc-2dac-460f-ac9e-90e288d1dd87', N'SubscriptionJobInterval', N'5', NULL, 0, NULL, NULL, NULL, NULL)
 INSERT [dbo].[IzendaSystemSetting] ([Id], [Name], [Value], [Version], [Deleted], [Created], [CreatedBy], [Modified], [ModifiedBy]) VALUES (N'8d721bfc-2dac-460f-ac9e-90e288d1dd89', N'ApiSecretToken', N'FF388608-8BE6-45E4-AC43-F11D6454161D', NULL, 0, NULL, NULL, NULL, NULL)
-INSERT [dbo].[IzendaSystemSetting] ([Id], [Name], [Value], [Version], [Deleted], [Created], [CreatedBy], [Modified], [ModifiedBy]) VALUES (N'1f9c13ca-fea4-467b-9a76-a17122821fe6', N'AuthRSAPublicKey', N'<RSAKeyValue><Modulus>rF0xAx3elTdRSt4Bk6c0wRalhjejr00XLIv3T8Wagcx9QqboU46+l+BXD1xmBm0WCmxjsZSpnjTAIAigWQ7EhPnZGD/kNHIuFnGW0HsorGA/E9+6FQXYAQMeF6iclI+JUKgsNQm8p8QRFG93pllMQalnePbdF/cR/5d0XXq8Jec=</Modulus><Exponent>AQAB</Exponent></RSAKeyValue>', NULL, 0, NULL, NULL, NULL, NULL)
+INSERT [dbo].[IzendaSystemSetting] ([Id], [Name], [Value], [Version], [Deleted], [Created], [CreatedBy], [Modified], [ModifiedBy]) VALUES (N'1f9c13ca-fea4-467b-9a76-a17122821fe6', N'AuthRSAPublicKey', N'<RSAKeyValue><Modulus>hC4mGSteilxxqFgsrHxprLpYUHEFaE+aTMSQ5fCf58x5LeEhIa92WMgX6oLdI2nss1OOZpTqhjfkCaiarUPPEHIriJa+aXeHNFByxQnzDkG1tNgMjXMF8Dynyqpv41fF24OTjIm2KVfc3KEX9f3wB3JuhHxw2zlx5aQ42TpFsvs=</Modulus><Exponent>AQAB</Exponent></RSAKeyValue>', NULL, 0, NULL, NULL, NULL, NULL)
 INSERT [dbo].[IzendaSystemSetting] ([Id], [Name], [Value], [Version], [Deleted], [Created], [CreatedBy], [Modified], [ModifiedBy]) VALUES (N'deb69225-e4a1-4533-9b29-a44f2eea8487', N'AllowDuplicateUser', N'1', NULL, 0, NULL, NULL, NULL, NULL)
 INSERT [dbo].[IzendaSystemSetting] ([Id], [Name], [Value], [Version], [Deleted], [Created], [CreatedBy], [Modified], [ModifiedBy]) VALUES (N'deb69225-e4a1-4533-9b29-a44f2eea8488', N'StoredProcParamDelimiter', N',', NULL, 0, NULL, NULL, NULL, NULL)
 INSERT [dbo].[IzendaSystemSetting] ([Id], [Name], [Value], [Version], [Deleted], [Created], [CreatedBy], [Modified], [ModifiedBy]) VALUES (N'2afc3bac-2daf-4487-9a90-a5c65ccd7a14', N'PageSize', N'A4', NULL, 0, NULL, NULL, NULL, NULL)
