+++ Mvc5StarterKit/Controllers/DashboardController.cs
@@ -2,8 +2,19 @@
 
 namespace Mvc5StarterKit.Controllers
 {
+    [Authorize]
     public class DashboardController : Controller
     {
+        public ActionResult Index()
+        {
+            return View();
+        }
+
+        public ActionResult DashboardDesigner()
+        {
+            return View();
+        }
+
         // GET: DashboardViewer
         public ActionResult DashboardViewer(string id)
         {
