+++ Mvc5StarterKit/Controllers/IzendaController.cs
@@ -6,24 +6,16 @@ using System.Web.Mvc;
 
 namespace Mvc5StarterKit.Controllers
 {
-    public class HomeController : Controller
+    [Authorize]
+    public class IzendaController : Controller
     {
         public ActionResult Index()
         {
             return View();
         }
 
-        public ActionResult About()
+        public ActionResult Settings()
         {
-            ViewBag.Message = "Your application description page.";
-
-            return View();
-        }
-
-        public ActionResult Contact()
-        {
-            ViewBag.Message = "Your contact page.";
-
             return View();
         }
     }
