--- Mvc5StarterKit/Models/UserInfo.cs
@@ -1,13 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Web;
-
-namespace Mvc5StarterKit.Models
-{
-    public class UserInfo
-    {
-        public string UserName { get; set; }
-        public string TenantUniqueName { get; set; }
-    }
-}
