+++ Mvc5StarterKit/Views/Home/Index.cshtml
@@ -1,5 +1,5 @@
 ﻿@{
-    ViewBag.Title = "Home Page";
+    ViewBag.Title = "Home";
 }
 <div class="jumbotron izenda-header">
     <h1>@*Izenda Integrated Application Demo*@</h1>
