--- Mvc5StarterKit/Views/Home/Izenda.cshtml
@@ -1,16 +1,0 @@
-﻿
-@{
-    ViewBag.Title = "Izenda";
-}
-
-
-@section scripts
-{
-    <script type="text/javascript">
-        $(document).ready(function () {
-            izendaInit();
-        });
-    </script>
-}
-<div class="loader" id="progressLoader"> </div>
-<div class="izenda-container" id="izenda-root"></div>
