--- Mvc5StarterKit/Views/Home/ReportDesigner.cshtml
@@ -1,15 +1,0 @@
-﻿@{
-    ViewBag.Title = "Report Designer";
-}
-
-@section scripts
-{
-    <script type="text/javascript">
-        $(document).ready(function () {
-            izendaInitReportDesigner();
-        });
-    </script>
-}
-
-<div class="loader" id="progressLoader"> </div>
-<div class="izenda-container" id="izenda-root"></div>
