+++ Mvc5StarterKit/Views/Izenda/Settings.cshtml
@@ -1,0 +1,15 @@
+﻿@{
+    ViewBag.Title = "Settings";
+}
+
+@section scripts
+{
+    <script type="text/javascript">
+        $(document).ready(function () {
+            izendaInitSetting();
+        });
+    </script>
+}
+
+<div class="loader" id="progressLoader"> </div>
+<div class="izenda-container" id="izenda-root"></div>
