+++ Mvc5StarterKit/Web.config
@@ -18,35 +18,31 @@
     <add key="ClientValidationEnabled" value="true" />
     <add key="UnobtrusiveJavaScriptEnabled" value="true" />
     <!-- RSA private key in PEM format -->
-    <add key="RSAPrivateKey" value="MIICXAIBAAKBgQDMVlBx0jp/bIiDibx+f5SeUY9rwqV/TVkgg8fVLZgUqYmk8hEZqzThvZEan4epXmOYHjDsKWugr2Si25YqDXJKpow62D6flwv+VuDHePctgUWxf1NlPCL4wFSk718HAFazZCXL99XOK/gVjSndhcEk6z+9lk1HmXCHD93KwXou9QIDAQABAoGACevO8BfS+0jbv/c6DbJIFv/CxOqoemvY/fkoBLO4BJjOtBGEvwhPAv7fQrmoLpMEpuggW/cO4LhjXHzo55XLjLoRjBBbiPbZayaAeptP9oYMyBNwBp9d49taawXm7nxiOC8sszkzJ0gKFeN+plTQruDm+HspaGBmUHdCMlJ9zakCQQD1cCZpxV1suxk4l2rXUa9edWmh5aYQj/nagtD/fBSkGcN6Q2+yk4y2V7wb/t0pHeab6p6eSu5MXGcOl9VXncbjAkEA1SFhB7AFT+/ehxDLgwdWEdBFRdkQzEbzNmk1lKgvZf8amipAw4n7DEjSoyqIXqXXr5DdyqSUDARylWnfzADCRwJABcsm7Po+sVFdUAuq9vgzpowo+Sxdlih/4luSKWW5awI8rgcnfNSkzq0VgKesesr85ZNNOTlVlLHdsOd+nrnXtwJARUqr3C77GykWRP1N3RS2g+Ydj37p+jAbBJaiB+nCNzwALx0Ln0ct6qmGaev7GCJ9BCRqJ2bohxuvESqxywZ4IwJBAM438RdcURHNUzY2xVVF9Ab/jGlGZahLMyx+tFgocS+wK1ZGuQzWz7GPPPVDL1rjNP88pnyFJSXMR2zFPXZm7jo=" />
+    <add key="RSAPrivateKey" value="MIICXAIBAAKBgQCELiYZK16KXHGoWCysfGmsulhQcQVoT5pMxJDl8J/nzHkt4SEhr3ZYyBfqgt0jaeyzU45mlOqGN+QJqJqtQ88QciuIlr5pd4c0UHLFCfMOQbW02AyNcwXwPKfKqm/jV8Xbg5OMibYpV9zcoRf1/fAHcm6EfHDbOXHlpDjZOkWy+wIDAQABAoGAAZx9NJq9o2RYf4vBOHV0/x9WAUQ6o4vI3tbmL0j9VTpXkp9No8ph4Ar8Y6YmicGtgLG4wpGRJ4OATglv06nhpclZAjFcb7e7lpr8lZYJQ2q88i3x+8z4/4Lv9hWBzwYuQx45SgZASqKHlRXDGILu2gvVAXqfwbRHm45bw9M9QaUCQQDsnaSOCTptSFp2lkk7Fm97mQPOaKeNSMNMY+3jBQMYV2WdfiVzPK91eSMFuyRPfxDYjVf3342CibKCmL4k/usNAkEAjwJDBbfh2xMhcgqzLHJYVy/5+QVs6b6vy40j7UvATaGiowNYr+V4FUUeRmuA9AVNSRRw27dMfDLRodOztSZ0JwJAENtmRsw3gmOeNkacrv35MUgEcmcYUMN3TnDsGLlZsBlB3weGqYAQTAE2UokznApEa/kd5EXj8gHXWuCA42fZ3QJBAIig25GaBbE9DMYczOPY9ueNkaBFe7GyGdg6JynuJnLEy1I8vz2F2X1QxMT4roH0kssQOWbhL7S1BK08SkP20h8CQEbD7vartA5WL74V8VNgGThBuPqnpsagTChHO0Rat5jR6nG2XHDAWiaFMJ/m/0sLjAAWl6s0gA6CTP/Yh6CPtBY=" />
 
     <!--Izenda-->
     <add key="IzendaAdminUser" value="IzendaAdmin@system.com" />
-    <add key="IzendaApiUrl" value="http://localhost:65469/api/" />
+    <add key="IzendaApiUrl" value="http://localhost:5101/api/" />
     <!--Izenda End-->
 
   </appSettings>
   <system.web>
-
     <authentication mode="None" />
     <compilation debug="true" targetFramework="4.5.2" />
     <httpRuntime targetFramework="4.5.2" />
-    <httpModules>
-    </httpModules>
   </system.web>
   <system.webServer>
     <modules>
       <remove name="FormsAuthentication" />
     </modules>
     <validation validateIntegratedModeConfiguration="false" />
-
-    
-  <handlers>
+    <handlers>
       <remove name="ExtensionlessUrlHandler-Integrated-4.0" />
       <remove name="OPTIONSVerbHandler" />
       <remove name="TRACEVerbHandler" />
       <add name="ExtensionlessUrlHandler-Integrated-4.0" path="*." verb="*" type="System.Web.Handlers.TransferRequestHandler" preCondition="integratedMode,runtimeVersionv4.0" />
-    </handlers></system.webServer>
+    </handlers>
+  </system.webServer>
   <runtime>
     <assemblyBinding xmlns="urn:schemas-microsoft-com:asm.v1">
       <dependentAssembly>
@@ -63,7 +59,7 @@
       </dependentAssembly>
       <dependentAssembly>
         <assemblyIdentity name="Microsoft.Owin" publicKeyToken="31bf3856ad364e35" />
-        <bindingRedirect oldVersion="0.0.0.0-3.0.1.0" newVersion="3.0.1.0" />
+        <bindingRedirect oldVersion="0.0.0.0-4.0.0.0" newVersion="4.0.0.0" />
       </dependentAssembly>
       <dependentAssembly>
         <assemblyIdentity name="Newtonsoft.Json" culture="neutral" publicKeyToken="30ad4fe6b2a6aeed" />
