+++ Mvc5StarterKit/packages.config
@@ -5,6 +5,7 @@
   <package id="EntityFramework" version="6.1.3" targetFramework="net452" />
   <package id="jQuery" version="1.10.2" targetFramework="net452" />
   <package id="jQuery.Validation" version="1.11.1" targetFramework="net452" />
+  <package id="Microsoft.AspNet.Cors" version="5.0.0" targetFramework="net452" />
   <package id="Microsoft.AspNet.Identity.Core" version="2.2.1" targetFramework="net452" />
   <package id="Microsoft.AspNet.Identity.EntityFramework" version="2.2.1" targetFramework="net452" />
   <package id="Microsoft.AspNet.Identity.Owin" version="2.2.1" targetFramework="net452" />
@@ -20,7 +21,8 @@
   <package id="Microsoft.CodeDom.Providers.DotNetCompilerPlatform" version="1.0.0" targetFramework="net452" />
   <package id="Microsoft.jQuery.Unobtrusive.Validation" version="3.2.3" targetFramework="net452" />
   <package id="Microsoft.Net.Compilers" version="1.0.0" targetFramework="net452" developmentDependency="true" />
-  <package id="Microsoft.Owin" version="3.0.1" targetFramework="net452" />
+  <package id="Microsoft.Owin" version="4.0.0" targetFramework="net452" />
+  <package id="Microsoft.Owin.Cors" version="4.0.0" targetFramework="net452" />
   <package id="Microsoft.Owin.Host.SystemWeb" version="3.0.1" targetFramework="net452" />
   <package id="Microsoft.Owin.Security" version="3.0.1" targetFramework="net452" />
   <package id="Microsoft.Owin.Security.Cookies" version="3.0.1" targetFramework="net452" />
